Restructure Canvas replay classes
